critical: In the dark mode, you can not see the text in the log and in other parts of the form fields. This needs to be fixed for the dark mode so it's easy to read and see. The icons are bright enough but the text as you can see from this screenshot is not.
86a23ajg7

critical: In the dark mode, you can not see the text in the log and in other parts of the form fields. This needs to be fixed for the dark mode so it's easy to read and see. The icons are bright enough but the text as you can see from this screenshot is not.